*: refactor Dockerfiles and Makefiles (#4069)
