+++ images/operator-sdk/Dockerfile
@@ -1,7 +1,8 @@
 FROM registry.access.redhat.com/ubi8/ubi-minimal:latest
 
-COPY operator-sdk /usr/local/bin/operator-sdk
-
 RUN microdnf install -y golang make which
+
+ARG BIN=operator-sdk
+COPY $BIN /usr/local/bin/operator-sdk
 
 ENTRYPOINT ["/usr/local/bin/operator-sdk"]
